Move middleware to own file.
